MNT/work-requisition/Error message on delete constraint

+++ backend/Modules/Maintenance/Http/Controllers/MntItemController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Modules\Maintenance\Entities\MntItem;
 use Modules\Maintenance\Entities\MntItemGroup;
+use Modules\Maintenance\Entities\MntJob;
 use Modules\Maintenance\Http\Requests\MntItemRequest;
 use Modules\Maintenance\Http\Requests\MntItemUpdateRequest;
 
@@ -127,7 +128,17 @@ class MntItemController extends Controller
      */
     public function destroy($id)
     {
-        try {            
+        try {          
+            $jobs = MntJob::where('mnt_item_id', $id)->count();
+            if ($items > 0) {
+                $error = array(
+                    "message" => "Data could not be deleted!",
+                    "errors" => [
+                        "id"=>"This data could not be deleted as it has reference to other table"
+                    ]
+                );
+                return response()->json($error, 422);
+            }
             $item = MntItem::findorfail($id);
             $item->delete();
